revert commit 4028149  review 571170 (last sync yt to arcadia)
git-svn-id: svn+ssh://arcadia.yandex.ru/arc/trunk/arcadia/yt/19_4/yt@4028376 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ ytlib/query_client/query_preparer.cpp
@@ -1933,8 +1933,7 @@ public:
             if (Mapping_) {
                 Mapping_->push_back(TColumnDescriptor{
                     formattedName,
-                    SourceTableSchema_.GetColumnIndex(*column)
-                });
+                    size_t(SourceTableSchema_.GetColumnIndex(*column))});
             }
 
             return TBaseColumn(formattedName, column->GetPhysicalType());
